requirements.txt
absl-py==1.2.0
accelerate==0.12.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.2.1
alembic==1.14.0
async-timeout==4.0.2
attrs==22.1.0
blinker==1.8.2
boto3==1.35.88
botocore==1.35.88
cachetools==5.2.0
catalyst==22.4
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.8
cloudpickle==3.1.0
cycler==0.11.0
databricks-sdk==0.40.0
Deprecated==1.2.15
docker==7.1.0
efficientnet-pytorch==0.7.1
flask==3.0.3
fonttools==4.36.0
frozenlist==1.3.1
fsspec==2022.7.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.10.0
google-auth-oauthlib==0.4.6
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
greenlet==3.1.1
grpcio==1.47.0
gunicorn==23.0.0
hydra-slayer==0.4.0
idna==3.3
imageio==2.21.1
imgaug==0.4.0
importlib-metadata==4.12.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.5
jmespath==1.0.1
joblib==1.1.0
kiwisolver==1.4.4
lmdb==1.3.0
Mako==1.3.8
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
mlflow==2.17.2
mlflow-skinny==2.17.2
multidict==6.0.2
munch==2.5.0
networkx==2.8.5
numpy==1.23.2
oauthlib==3.2.0
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
pretrainedmodels==0.7.4
protobuf==3.20.1
psutil==5.9.1
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.7.2
pytz==2022.2.1
PyWavelets==1.3.0
PyYAML==6.0
qudida==0.0.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.10.4
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.0
seaborn==0.11.2
Shapely==1.8.4
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.36
sqlparse==0.5.3
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
threadpoolctl==3.1.0
tifffile==2022.8.12
torch==1.8.1+cu111
torchaudio==0.8.1
torchmetrics==0.9.3
torchvision==0.9.1+cu111
tqdm==4.64.0
typing-extensions==4.12.2
urllib3==1.26.11
werkzeug==3.0.6
wrapt==1.17.0
yarl==1.8.1
zipp==3.8.1
